Review fix, move guided commands to deployment
