+++ alertlogiccli/commands/deployment/__init__.py
@@ -2,6 +2,9 @@ from . import mode
 from . import status
 from . import resources
 from . import scan_queue
+from . import subnet
+from . import installation
+from . import scanner
 
 metadata = {
     "subcommands": [
@@ -10,7 +13,12 @@ metadata = {
         resources.ListDeployed(),
         status.GetStatus(),
         scan_queue.ScanHost(),
-        scan_queue.ListScanQueues()
+        scan_queue.ListScanQueues(),
+        subnet.SetSubnet(),
+        subnet.GetConfiguration(),
+        installation.InstallationStatus(),
+        installation.Redeploy(),
+        scanner.ScannerEstimation()
     ],
     "name": "deployment",
     "help": "deployment related commands"
