+++ alertlogiccli/commands/deployment/scanner.py
