+++ tests/installation_tests.py
@@ -1,11 +1,9 @@
-import pytest
-import mock
 import httpretty
-
-import mock_tools
+import pytest
 
 import alertlogiccli.command
-import alertlogiccli.commands.guided.installation
+import alertlogiccli.commands.deployment.installation
+import mock_tools
 
 
 class TestInstallationStatus():
@@ -21,7 +19,7 @@ class TestInstallationStatus():
         )
         with pytest.raises(alertlogiccli.command.InvalidHTTPResponse):
             context = mock_tools.make_context(args)
-            alertlogiccli.commands.guided.installation.InstallationStatus().execute(context)
+            alertlogiccli.commands.deployment.installation.InstallationStatus().execute(context)
 
     @httpretty.activate
     def test_server_error(self):
@@ -34,7 +32,7 @@ class TestInstallationStatus():
         )
         with pytest.raises(alertlogiccli.command.InvalidHTTPResponse):
             context = mock_tools.make_context(args)
-            alertlogiccli.commands.guided.installation.InstallationStatus().execute(context)
+            alertlogiccli.commands.deployment.installation.InstallationStatus().execute(context)
 
     @httpretty.activate
     def test_created(self):
@@ -54,7 +52,7 @@ class TestInstallationStatus():
             content_type="text/json"
         )
         context = mock_tools.make_context(args)
-        result = alertlogiccli.commands.guided.installation.InstallationStatus().execute(context)
+        result = alertlogiccli.commands.deployment.installation.InstallationStatus().execute(context)
         assert (result == '['
                           '{'
                           '"account_id":"010000001",'
@@ -76,7 +74,7 @@ class TestRedeploy():
         )
         with pytest.raises(alertlogiccli.command.InvalidHTTPResponse):
             context = mock_tools.make_context(args)
-            alertlogiccli.commands.guided.installation.Redeploy().execute(context)
+            alertlogiccli.commands.deployment.installation.Redeploy().execute(context)
 
     @httpretty.activate
     def test_server_error(self):
@@ -89,7 +87,7 @@ class TestRedeploy():
         )
         with pytest.raises(alertlogiccli.command.InvalidHTTPResponse):
             context = mock_tools.make_context(args)
-            alertlogiccli.commands.guided.installation.Redeploy().execute(context)
+            alertlogiccli.commands.deployment.installation.Redeploy().execute(context)
 
     @httpretty.activate
     def test_created(self):
@@ -101,5 +99,5 @@ class TestRedeploy():
             status=204,
         )
         context = mock_tools.make_context(args)
-        result = alertlogiccli.commands.guided.installation.Redeploy().execute(context)
+        result = alertlogiccli.commands.deployment.installation.Redeploy().execute(context)
         assert (result == "ok")
